Add and use import pipeline for sozialpass nuernberg
