+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/pipelines/Pipeline.kt
@@ -1,0 +1,8 @@
+package app.ehrenamtskarte.backend.stores.importer.pipelines
+
+import app.ehrenamtskarte.backend.config.BackendConfiguration
+import org.slf4j.Logger
+
+interface Pipeline {
+    fun import(config: BackendConfiguration, logger: Logger)
+}
