+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/pipelines/SozialpassNuernberg.kt
@@ -1,0 +1,10 @@
+package app.ehrenamtskarte.backend.stores.importer.pipelines
+
+import app.ehrenamtskarte.backend.config.BackendConfiguration
+import org.slf4j.Logger
+
+object SozialpassNuernberg : Pipeline {
+    override fun import(config: BackendConfiguration, logger: Logger) {
+        logger.info("Not implemented yet!")
+    }
+}
