+++ backend/src/main/resources/config/config.yml
@@ -13,4 +13,7 @@ server:
 projects:
   - id: bayern.ehrenamtskarte.app
     importUrl: https://www.lbe.bayern.de/engagement-anerkennen/ehrenamtskarte/akzeptanzstellen/app-daten.xml
-    pipelineName: lbe-bayern
+    pipelineName: EhrenamtskarteBayern
+  - id: nuernberg.sozialpass.app
+    importUrl: https://example.com
+    pipelineName: SozialpassNuernberg
